+++ src/main/java/com/simsasookbak/external/ai/alan/controller/AlanController.java
@@ -21,4 +21,5 @@ public class AlanController {
         AlanResponseDto response = alanService.getAlan(content);
         return ResponseEntity.ok(response);
     }
+
 }
